test(architectures): 拆分优先级测试辅助类型文件
- 拆分 PriorityServiceTests 末尾的测试接口与辅助类型到独立文件
- 保留原有命名空间、可见性与优先级测试语义
- 补充新文件的 XML 注释并清理原测试文件的无关 using
