+++ GFramework.Core.Tests/Architectures/IMixedTestSystem.cs
@@ -1,0 +1,10 @@
+using GFramework.Core.Abstractions.Systems;
+
+namespace GFramework.Core.Tests.Architectures;
+
+/// <summary>
+///     定义用于混合优先级排序测试的系统契约。
+/// </summary>
+public interface IMixedTestSystem : ISystem
+{
+}
