+++ GFramework.Core.Tests/Architectures/IPriorityTestModel.cs
@@ -1,0 +1,10 @@
+using GFramework.Core.Abstractions.Model;
+
+namespace GFramework.Core.Tests.Architectures;
+
+/// <summary>
+///     定义用于优先级排序测试的模型契约。
+/// </summary>
+public interface IPriorityTestModel : IModel
+{
+}
