+++ GFramework.Core.Tests/Architectures/IPriorityTestSystem.cs
@@ -1,0 +1,10 @@
+using GFramework.Core.Abstractions.Systems;
+
+namespace GFramework.Core.Tests.Architectures;
+
+/// <summary>
+///     定义用于优先级排序测试的系统契约。
+/// </summary>
+public interface IPriorityTestSystem : ISystem
+{
+}
