+++ GFramework.Core.Tests/Architectures/IPriorityTestUtility.cs
@@ -1,0 +1,10 @@
+using GFramework.Core.Abstractions.Utility;
+
+namespace GFramework.Core.Tests.Architectures;
+
+/// <summary>
+///     定义用于优先级排序测试的工具契约。
+/// </summary>
+public interface IPriorityTestUtility : IUtility
+{
+}
