+++ GFramework.Core.Tests/Architectures/PriorityServiceTests.cs
@@ -1,12 +1,7 @@
 using System.Reflection;
-using GFramework.Core.Abstractions.Bases;
 using GFramework.Core.Abstractions.Logging;
-using GFramework.Core.Abstractions.Model;
-using GFramework.Core.Abstractions.Systems;
-using GFramework.Core.Abstractions.Utility;
 using GFramework.Core.Ioc;
 using GFramework.Core.Logging;
-using GFramework.Core.Model;
 
 namespace GFramework.Core.Tests.Architectures;
 
@@ -120,129 +115,3 @@ public class PriorityServiceTests
         Assert.That(systems[2], Is.InstanceOf<MixedTestSystemWithPriority>()); // 10
     }
 }
-
-#region Test Interfaces
-
-public interface IPriorityTestSystem : ISystem
-{
-}
-
-public interface IPriorityTestModel : IModel
-{
-}
-
-public interface IPriorityTestUtility : IUtility
-{
-}
-
-public interface IMixedTestSystem : ISystem
-{
-}
-
-#endregion
-
-#region Test Systems
-
-public class PriorityTestSystemA : AbstractSystem, IPriorityTestSystem, IPrioritized
-{
-    public int Priority => 10;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class PriorityTestSystemB : AbstractSystem, IPriorityTestSystem, IPrioritized
-{
-    public int Priority => 20;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class PriorityTestSystemC : AbstractSystem, IPriorityTestSystem, IPrioritized
-{
-    public int Priority => 30;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class MixedTestSystemWithPriority : AbstractSystem, IMixedTestSystem, IPrioritized
-{
-    public int Priority => 10;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class MixedTestSystemWithoutPriority : AbstractSystem, IMixedTestSystem
-{
-    protected override void OnInit()
-    {
-    }
-}
-
-public class MixedTestSystemNegativePriority : AbstractSystem, IMixedTestSystem, IPrioritized
-{
-    public int Priority => -10;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-#endregion
-
-#region Test Models
-
-public class PriorityTestModelA : AbstractModel, IPriorityTestModel, IPrioritized
-{
-    public int Priority => 10;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class PriorityTestModelB : AbstractModel, IPriorityTestModel, IPrioritized
-{
-    public int Priority => 20;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-public class PriorityTestModelC : AbstractModel, IPriorityTestModel, IPrioritized
-{
-    public int Priority => 30;
-
-    protected override void OnInit()
-    {
-    }
-}
-
-#endregion
-
-#region Test Utilities
-
-public class PriorityTestUtilityA : IPriorityTestUtility, IPrioritized
-{
-    public int Priority => 10;
-}
-
-public class PriorityTestUtilityB : IPriorityTestUtility, IPrioritized
-{
-    public int Priority => 20;
-}
-
-public class PriorityTestUtilityC : IPriorityTestUtility, IPrioritized
-{
-    public int Priority => 30;
-}
-
-#endregion
